requirements.txt
opencv-python
pynput
plyer
pystray
Pillow
deepface
mediapipe 
scipy
tf-keras